+++ inc/admin/welcome-screen/sections/getting-started.php
@@ -5,10 +5,9 @@
 ?>
 <?php
 // get theme customizer url
-$url 	= admin_url() . 'customize.php?';
-$url 	.= 'url=' . urlencode( site_url() . '?storefront-customizer=true' );
-$url 	.= '&return=' . urlencode( admin_url() . 'themes.php?page=storefront-welcome' );
-$url 	.= '&storefront-customizer=true';
+$customizer_url 	= admin_url() . 'customize.php?url=' . urlencode( site_url() . '?storefront-customizer=true' ) . '&return=' . urlencode( admin_url() . 'themes.php?page=storefront-welcome' ) . '&storefront-customizer=true';
+$menus_url 			= admin_url() . 'customize.php?autofocus%5Bsection%5D=nav&?url=' . urlencode( site_url() . '?storefront-customizer=true' ) . '&return=' . urlencode( admin_url() . 'themes.php?page=storefront-welcome' ) . '&storefront-customizer=true';
+
 ?>
 <div id="getting_started" class="col panel" style="margin-bottom: 1.618em; padding-top: 1.618em; overflow: hidden;">
 
@@ -34,13 +33,13 @@ $url 	.= '&storefront-customizer=true';
 	<p><?php esc_html_e( 'The secondary navigation is better suited to lower traffic pages such as terms and conditions.', 'storefront' ); ?></p>
 	<p><?php esc_html_e( 'The handheld navigation gives you complete control over what menu items to serve to handheld devices.', 'storefront' ); ?></p>
 
-	<p><a href="<?php echo esc_url( self_admin_url( 'nav-menus.php' ) ); ?>" class="button"><?php esc_html_e( 'Configure menus', 'storefront' ); ?></a></p>
+	<p><a href="<?php echo esc_url( $menus_url ); ?>" class="button"><?php esc_html_e( 'Configure menus', 'storefront' ); ?></a></p>
 
 	<hr />
 
 	<h4><?php esc_html_e( 'Create a color scheme' ,'storefront' ); ?></h4>
 	<p><?php esc_html_e( 'Using the WordPress Customizer you can tweak Storefront\'s appearance to match your brand.', 'storefront' ); ?></p>
-	<p><a href="<?php echo esc_url( $url ); ?>" class="button"><?php esc_html_e( 'Open the Customizer', 'storefront' ); ?></a></p>
+	<p><a href="<?php echo esc_url( $customizer_url ); ?>" class="button"><?php esc_html_e( 'Open the Customizer', 'storefront' ); ?></a></p>
 
 	<hr />